refactor: expose UI helpers under gpui_kit::test
